// app/types/object.ts
export interface Book {
    id: number;
    user_id: string;
    isbn: string;
    title: string;
    subtitle: string;
    author: string;
    published: string;
    publisher: string;
    pages: number;
    description: string;
    website: string;
    created_at: string;
    updated_at: string;
};